# backend/internal/dashboard/data/openapi.json
{
  "components": {
    "schemas": {
...
      "PageStateRequest": {
        "properties": {
          "data": {
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "PageStateResponse": {
        "properties": {
          "data": {}
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "RestartResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
      "ResultEnvelope": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "fb_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard admin surface. Generated by backend/cmd/openapi-emit from admin_manifest.json + dashboard.AdminAPIPaths — do not edit by hand.",
    "title": "freebucks-proxy admin API",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/api/pages/{id}": {
      "get": {
        "operationId": "getPageState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageStateResponse"
                }
              }
            },
            "description": "Per-page UI snapshot (absent row reads back empty, never 404)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Per-page UI snapshot (absent row reads back empty, never 404)",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "putPageState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            },
            "description": "Upsert one per-page UI snapshot (data must be a JSON object)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Upsert one per-page UI snapshot (data must be a JSON object)",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/admin/restart": {
      "post": {
        "operationId": "restart",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestartResponse"
                }
              }
            },
            "description": "Validate config and restart the gateway process"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Validate config and restart the gateway process",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}